Print the first 50 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,15 +9,15 @@
 int main(void)
 {
 	int n = 50, i;
-	int fib1 = 1, fib2 = 2, next;
+	long long int fib1 = 1, fib2 = 2, next;
 
-	printf("%d, %d, ", fib1, fib2);
+	printf("%lld, %lld, ", fib1, fib2);
 
 	for (i = 3; i <= n; i++)
 	{
 		next = fib1 + fib2;
 
-		printf("%d, ", next);
+		printf("%lld, ", next);
 		fib1 = fib2;
 		fib2 = next;
 	}
